CAS LOGIN SYSTEM (NOT FINISHED)
Right now, I implemented part of CAS authentication, however, because our project is not yet 'whitelisted', CAS will not work 100% correctly and will not send back information to our project. To fix this, I created a DEV login button just so we can login ourselves (albiet unsecured) for testing purposes. We will remove this dev login for the final project.